requirements.txt
# requirements.txt

fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
google-generativeai
pinecone
PyMuPDF
langchain
gunicorn